Add metadata properties to CycloneDX
This commit adds arbitrary properties from the syft metadata to
CycloneDX. This builds on top of the previous work on enriching
the cyclonedx output. Creating a separate PR for this work as this
might be a point of contention whether we want to add these properties
or not or the way the properties are currently extracted using struct
tags.

Signed-off-by: Sambhav Kothari <skothari44@bloomberg.net>