Normalize syft-json output
There are numerous places within syft where a map is iterated over to
create a list. Since map iteration is non-deterministic, the order of
items within the list is also non-deterministic. While this doesn't
impact the semantic meaning of the output, it makes it harder to compare
the output of two independent runs of syft.

```sh
syft ghcr.io/jenkins-x/jx-scm -o syft-json > a.json
syft ghcr.io/jenkins-x/jx-scm -o syft-json > b.json
diff a.json b.json
```

Without this PR there is a diff, with this PR the diff is empty.

This PR is a band-aid to address the most common sources of drift. There
are certainly other sources that are not addressed.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>