feat: add identifiable field to source object
Allow source.Source struct to set reproducable id for different scheme.
This ID is calcuated either as a digest from the given directory or file
path. If the scheme is detected to be an ImageScheme then the ID is
calculated as a ChainID: https://github.com/opencontainers/image-spec/blob/main/config.md#layer-chainid

Signed-off-by: Christopher Phillips <christopher.phillips@anchore.com>