chore: regenerate snapshot fixtures for all format
This PR introduces a hash:ignore tag for the virtual path in Java
metadata. It also introduces a hash:ignore for the Locations field of a
given package. Because these two are no longer considered as part of the
struct hash golden snapshots needed to be regenerated. No other data
should have changed beyond the ID/related fields or timestamps

Signed-off-by: Christopher Phillips <christopher.phillips@anchore.com>