fix: add virtualPath back to hash struct method
virtualPath needs to be considered as a part of ID generation for
packages that use the JavaMetadata struct. Consider a package that has a
virtualPath of "/java/example-jenkins-plugin.hpi" vs one that has the
virtualPath of "/java/example-jenkins-plugin.hpi:WEB-INF/lib/example-jenkins-plugin.jar".
Syft views these as separate packages and thus they should be cataloged
and sorted separately and in a stable fashion. Conversely, if an apkdb
package is found to have exactly the same metadata, it is considered the
same package and the location lists are to be merged.

Signed-off-by: Christopher Phillips <christopher.phillips@anchore.com>