source: plumb sym link base through to syft CLI
Continuing the work of b81c980, change the behavior of the syft
command: the root for symbol link resolution of directory scans is now
the directory root, not "/".

Do this pedantically, by separating the directory root to the scanned
from the directory sym link base.  As of this commit, these two values
will always be the same, but this strategy leaves the door open for
them to be different someday, if that would ever be useful.

Also remove some unused functions and consolidate down to just the
NewFromDirectoryBaseWithName() function.

Signed-off-by: Jeff Squyres <jeff@squyres.com>