Support scanning directories containing absolute symlinks in a chroot
Although syft supports scanning an entire chroot, it couldn't scan paths
inside the chroot containing absolute symlinks relative to the chroot
base until now.

This is the first step in supporting scanning files/directories in other
mount namespaces, since a different mount namespace is similar to a
chroot.

The main change is the addition to EvalSymlinksRelativeToBase, which can
resolve both regular symlinks and absolute symlinks confined in a
chroot-like directory structure. This function returns a path that can
be accessed in syft's scanning context, i.e. not meant to be used from
within the chroot context.

See #3396 for further discussions.

Signed-off-by: Ariel Miculas-Trif <amiculas@cisco.com>